Entropy exponent alpha should be an integer, not a double. Document values.
git-svn-id: https://svn.aspect.dealii.org/trunk/aspect@1606 244a8438-9d6d-46c2-a098-b74c898f6b90

+++ include/aspect/simulator.h
@@ -208,7 +208,7 @@ namespace aspect
          * @name Parameters that have to do with the stabilization of transport equations
          * @{
          */
-        double                         stabilization_alpha;
+        unsigned int                   stabilization_alpha;
         double                         stabilization_c_R;
         double                         stabilization_beta;
         /**
